Removed loop
Whatever happened on the 7th added a function somewhere that appears to
run on an interval and resizes the progress bar, so I can't edit it with
styles and adding a fast interval was wonky.  Going to need to figure
out how to edit or override a running function, somewhere in youtubes
minified scripts, OR find out what its referencing and edit that but no
luck yet.  Ew.