Center progressbar, fix related videos
Problem from last month remains, I need to find where the width is
grabbed in JS and edit that but I spent a good while searching the stack
and had no luck.

So, just centered the progess bar with css injected into the head that
changes a class, and fixed a problem with the related videos that
recently began appearing.